+++ src/Service/Firehose/src/FirehoseClient.php
@@ -28,6 +28,11 @@ class FirehoseClient extends AbstractApi
      * delivery stream. For more information about limits and how to request an increase, see Amazon Kinesis Data Firehose
      * Limits [^1].
      *
+     * Kinesis Data Firehose accumulates and publishes a particular metric for a customer account in one minute intervals.
+     * It is possible that the bursts of incoming bytes/records ingested to a delivery stream last only for a few seconds.
+     * Due to this, the actual spikes in the traffic might not be fully visible in the customer's 1 minute CloudWatch
+     * metrics.
+     *
      * You must specify the name of the delivery stream and the data record when using PutRecord. The data record consists
      * of a data blob that can be up to 1,000 KiB in size, and any kind of data. For example, it can be a segment from a log
      * file, geographic location data, website clickstream data, and so on.
@@ -40,8 +45,12 @@ class FirehoseClient extends AbstractApi
      * The `PutRecord` operation returns a `RecordId`, which is a unique string assigned to each record. Producer
      * applications can use this ID for purposes such as auditability and investigation.
      *
-     * If the `PutRecord` operation throws a `ServiceUnavailableException`, back off and retry. If the exception persists,
-     * it is possible that the throughput limits have been exceeded for the delivery stream.
+     * If the `PutRecord` operation throws a `ServiceUnavailableException`, the API is automatically reinvoked (retried) 3
+     * times. If the exception persists, it is possible that the throughput limits have been exceeded for the delivery
+     * stream.
+     *
+     * Re-invoking the Put API operations (for example, PutRecord and PutRecordBatch) can result in data duplicates. For
+     * larger data assets, allow for a longer time out before retrying Put API operations.
      *
      * Data records sent to Kinesis Data Firehose are stored for 24 hours from the time they are added to a delivery stream
      * as it tries to send the records to the destination. If the destination is unreachable for more than 24 hours, the
@@ -84,6 +93,11 @@ class FirehoseClient extends AbstractApi
      * producer than when writing single records. To write single data records into a delivery stream, use PutRecord.
      * Applications using these operations are referred to as producers.
      *
+     * Kinesis Data Firehose accumulates and publishes a particular metric for a customer account in one minute intervals.
+     * It is possible that the bursts of incoming bytes/records ingested to a delivery stream last only for a few seconds.
+     * Due to this, the actual spikes in the traffic might not be fully visible in the customer's 1 minute CloudWatch
+     * metrics.
+     *
      * For information about service quota, see Amazon Kinesis Data Firehose Quota [^1].
      *
      * Each PutRecordBatch request supports up to 500 records. Each record in the request can be as large as 1,000 KB
@@ -117,8 +131,11 @@ class FirehoseClient extends AbstractApi
      * processing. This minimizes the possible duplicate records and also reduces the total bytes sent (and corresponding
      * charges). We recommend that you handle any duplicates at the destination.
      *
-     * If PutRecordBatch throws `ServiceUnavailableException`, back off and retry. If the exception persists, it is possible
-     * that the throughput limits have been exceeded for the delivery stream.
+     * If PutRecordBatch throws `ServiceUnavailableException`, the API is automatically reinvoked (retried) 3 times. If the
+     * exception persists, it is possible that the throughput limits have been exceeded for the delivery stream.
+     *
+     * Re-invoking the Put API operations (for example, PutRecord and PutRecordBatch) can result in data duplicates. For
+     * larger data assets, allow for a longer time out before retrying Put API operations.
      *
      * Data records sent to Kinesis Data Firehose are stored for 24 hours from the time they are added to a delivery stream
      * as it attempts to send the records to the destination. If the destination is unreachable for more than 24 hours, the