+++ src/Service/Ses/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## NOT RELEASED
+
 ## 1.6.0
 
 ### Added